finished most areas with loading screens

+++ backend/src/Api/Controllers/Classes/ClassesController.cs
@@ -26,6 +26,9 @@ public class ClassesController : ControllerBase
     [ProducesResponseType(typeof(string), 400)]
     public async Task<IActionResult> DeleteClass([FromRoute] Guid classGuid)
     {
+        var currentPerson = await _foundationQueries.Service.RecogniseCurrentPersonByRelatedPerson(teachersGuids.First());
+        if (!await _foundationPermissions.Service.CanManageClass(classGuid, currentPerson.Response))
+            return new StatusResponse("Forbidden");
         var resp = await _foundationCommands.Service.DeleteClass(classGuid);
         return resp.Status ? Ok() : BadRequest(resp.Message);
     }
